Organize  Java dir
Fix Fibonacci Series Code
