--- Program's_Contributed_By_Contributors/Java_Programs/hello.java
@@ -1,8 +1,0 @@
-import java.util.*;
-public class first
-{
-  public static void main(String args[])
-  {
-  System.out.println("Hello World");
-  }
-}
